небольшие исправления
