+++ app/src/main/java/ru/practicum/android/diploma/search/data/dto/Response.kt
@@ -5,6 +5,6 @@ const val RESULT_CODE_SUCCESS = 200
 const val RESULT_CODE_BAD_REQUEST = 400
 const val RESULT_CODE_SERVER_ERROR = 500
 
-open class Resp {
+open class Response {
     var resultCode = 0
 }
