+++ app/src/main/java/ru/practicum/android/diploma/search/data/dto/SearchResponse.kt
@@ -1,3 +1,3 @@
 package ru.practicum.android.diploma.search.data.dto
 
-data class SearchResponse(val items: List<VacancyDto>, val found: Int) : Resp()
+data class SearchResponse(val items: List<VacancyDto>, val found: Int) : Response()
