+++ app/src/main/java/ru/practicum/android/diploma/search/data/dto/components/Phone.kt
@@ -1,0 +1,8 @@
+package ru.practicum.android.diploma.search.data.dto.components
+
+data class Phone(
+    val city: String,
+    val comment: Any?,
+    val country: String,
+    val number: String,
+)
