+++ app/src/main/java/ru/practicum/android/diploma/search/data/network/NetworkClient.kt
@@ -1,7 +1,7 @@
 package ru.practicum.android.diploma.search.data.network
 
-import ru.practicum.android.diploma.search.data.dto.Resp
+import ru.practicum.android.diploma.search.data.dto.Response
 
 interface NetworkClient {
-    suspend fun doRequest(dto: Any): Resp
+    suspend fun doRequest(dto: Any): Response
 }
