+++ app/src/main/java/ru/practicum/android/diploma/search/data/network/RetrofitClient.kt
@@ -8,7 +8,7 @@ import ru.practicum.android.diploma.search.data.dto.RESULT_CODE_BAD_REQUEST
 import ru.practicum.android.diploma.search.data.dto.RESULT_CODE_NO_INTERNET
 import ru.practicum.android.diploma.search.data.dto.RESULT_CODE_SERVER_ERROR
 import ru.practicum.android.diploma.search.data.dto.RESULT_CODE_SUCCESS
-import ru.practicum.android.diploma.search.data.dto.Resp
+import ru.practicum.android.diploma.search.data.dto.Response
 import ru.practicum.android.diploma.search.data.dto.SearchRequest
 import ru.practicum.android.diploma.utils.isInternetAvailable
 import ru.practicum.android.diploma.vacancy.data.dto.VacancyRequest
@@ -18,15 +18,15 @@ class RetrofitClient(
     private val context: Context
 ) : NetworkClient {
 
-    override suspend fun doRequest(dto: Any): Resp {
+    override suspend fun doRequest(dto: Any): Response {
         if (!isInternetAvailable(context)) {
-            return Resp().apply { resultCode = RESULT_CODE_NO_INTERNET }
+            return Response().apply { resultCode = RESULT_CODE_NO_INTERNET }
 
         }
         return getResp(dto = dto)
     }
 
-    private suspend fun getResp(dto: Any): Resp {
+    private suspend fun getResp(dto: Any): Response {
         return withContext(Dispatchers.IO) {
             try {
                 when (dto) {
@@ -41,13 +41,13 @@ class RetrofitClient(
                     }
 
                     else -> {
-                        Resp().apply { resultCode = RESULT_CODE_BAD_REQUEST }
+                        Response().apply { resultCode = RESULT_CODE_BAD_REQUEST }
                     }
                 }
 
             } catch (e: HttpException) {
                 println("RetrofitClient error code: ${e.code()} message: ${e.message}")
-                Resp().apply { resultCode = RESULT_CODE_SERVER_ERROR }
+                Response().apply { resultCode = RESULT_CODE_SERVER_ERROR }
             }
         }
     }
