+++ app/src/main/res/values/strings.xml
@@ -51,4 +51,7 @@
     <string name="sector_title">Выбор отрасли</string>
     <string name="sector_enter_sector">Введите отрасль</string>
     <string name="sector_select">Выбрать</string>
+    <!-- strings for openEmail -->
+    <string name="email_subject">Тема письма</string>
+    <string name="email_text">Текст письма</string>
 </resources>
